moved location to shared
