--- src/modules/sample/resolvers/other.object.resolver.js
@@ -1,16 +1,0 @@
-const Query = {
-  otherObject: (parent, args, context) => ({
-    id: "other-object",
-    name: "Other Object",
-  }),
-};
-
-const OtherObject = {
-  id: (parent) => parent.id,
-  name: (parent) => parent.name,
-};
-
-export default {
-  Query,
-  OtherObject,
-};
