--- src/modules/sample/resolvers/some.object.resolver.js
@@ -1,18 +1,0 @@
-const Query = {
-  someObject: (parent, args, context) => ({
-    id: "some-object",
-    name: "Some Object",
-  }),
-};
-
-const SomeObject = {
-  id: (parent) => parent.id,
-  name: (parent) => parent.name,
-  description: (parent) =>
-    `Description of object with id: ${parent.id}; name: ${parent.name}`,
-};
-
-export default {
-  Query,
-  SomeObject,
-};
