--- src/modules/sample/types/other.object.graphql
@@ -1,8 +1,0 @@
-type OtherObject {
-  id: ID!
-  name: String!
-}
-
-type Query {
-  otherObject: OtherObject!
-}
