--- src/modules/sample/types/some.object.graphql
@@ -1,9 +1,0 @@
-type SomeObject {
-  id: ID!
-  name: String!
-  description: String!
-}
-
-type Query {
-  someObject: SomeObject!
-}
