+++ src/modules/shared/resolvers/index.js
