+++ src/modules/shared/resolvers/location.resolver.js
@@ -1,0 +1,9 @@
+const Location = {
+  address: (parent) => parent.address,
+  longitude: (parent) => parent.longitude,
+  latitude: (parent) => parent.latitude,
+};
+
+export default {
+  Location,
+};
